api/views.py
from django.http import HttpResponse
from rest_framework import viewsets

from .models import Todo
from .serializers import TodoSerializer


class TodoViewSet(viewsets.ModelViewSet):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer


def healthcheck(request):
    return HttpResponse("app running")
