pyproject.toml
[tool.poetry]
name = "migrations-sandbox"
version = "0.1.0"
description = ""
authors = ["zachvalenta <zjv5002@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.9"
djangorestframework = "^3.11.0"
django-extensions = "^3.0.3"
bpython = "^0.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
